refactor: update source_image_reference in main.tf
The source_image_reference block in the main.tf file has been refactored to use dynamic blocks. This change allows for more flexibility in handling the source image reference based on the value of the var.instance.source_image_id variable. The publisher, offer, sku, and version values are now set based on the var.instance.image object, with default values provided if the object is not present.